// src/generated/web/openapi/api.ts
export interface CapabilityOption {
    /**
     * 
     * @type {string}
     * @memberof CapabilityOption
     */
    'os'?: string;
    /**
     * 
     * @type {string}
     * @memberof CapabilityOption
     */
    'os_version'?: string;
    /**
     * 
     * @type {string}
     * @memberof CapabilityOption
     */
    'browser'?: string;
    /**
     * 
     * @type {string}
     * @memberof CapabilityOption
     */
    'browser_version'?: string;
    /**
     * 
     * @type {string}
     * @memberof CapabilityOption
     */
    'device'?: string;
    /**
     * 
     * @type {string}
     * @memberof CapabilityOption
     */
    'device_type'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof CapabilityOption
     */
    'unsupported'?: boolean;
}
/**
 * 
 * @export
 * @interface CreateUrlReplacementRequest
 */
export interface CreateUrlReplacementRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateUrlReplacementRequest
     */
    'pattern_url': string;
    /**
     * 
     * @type {string}
     * @memberof CreateUrlReplacementRequest
     */
    'replacement_url': string;
}